LayoutTests/fast/js/bitops-type-tag-expected.txt

Ensure various bit operators correctly tag the final result value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS aDouble>>27 is 0
PASS aDouble>>27|0 is 0
PASS aDouble>>0 is 100000000
PASS aDouble>>0|0 is 100000000
PASS aDouble|0 is 100000000
PASS (function(x){ return 2147483648>>>x; })(31) is 1
PASS successfullyParsed is true

TEST COMPLETE
